+++ pkg/provider/status_test.go
@@ -43,7 +43,6 @@ var _ = Describe("GetMachineStatus", func() {
 			Data: map[string][]byte{
 				"projectId":         []byte("11111111-2222-3333-4444-555555555555"),
 				"serviceAccountKey": []byte(`{"credentials":{"iss":"test"}}`),
-				"networkId":         []byte("770e8400-e29b-41d4-a716-446655440000"),
 			},
 		}
