+++ samples/secret.yaml
@@ -38,9 +38,3 @@ stringData:
   # STACKIT region (required by SDK v1.0.0+)
   # Valid regions: eu01-1, eu01-2, etc.
   region: "eu01-1"
-
-  # Cloud-init user data for VM bootstrapping
-  userData: |
-    #cloud-config
-    runcmd:
-      - echo "Machine bootstrapped"
