+++ test/e2e/e2e_affinity_test.go
@@ -32,11 +32,6 @@ var _ = Describe("MCM Provider STACKIT", func() {
     projectId: "12345678-1234-1234-1234-123456789012"
     serviceAccountKey: "{}"
     region: "eu01-1"
-    networkId: "770e8400-e29b-41d4-a716-446655440000"
-    userData: |
-      #cloud-config
-      runcmd:
-        - echo "Machine bootstrapped"
   `, secretName, testNamespace)
 			createAndTrackResource(ctx, "secret", secretName, testNamespace, secretYAML)
