+++ test/e2e/e2e_networking_test.go
@@ -31,10 +31,6 @@ stringData:
   projectId: "12345678-1234-1234-1234-123456789012"
   serviceAccountKey: "{}"
   region: "eu01-1"
-  userData: |
-    #cloud-config
-    runcmd:
-      - echo "Machine bootstrapped"
 `, secretName, testNamespace)
 			createAndTrackResource(ctx, "secret", secretName, testNamespace, secretYAML)
 
@@ -100,10 +96,6 @@ stringData:
   projectId: "12345678-1234-1234-1234-123456789012"
   serviceAccountKey: "{}"
   region: "eu01-1"
-  userData: |
-    #cloud-config
-    runcmd:
-      - echo "Machine bootstrapped"
 `, secretName, testNamespace)
 			createAndTrackResource(ctx, "secret", secretName, testNamespace, secretYAML)
 
@@ -171,10 +163,6 @@ stringData:
   projectId: "12345678-1234-1234-1234-123456789012"
   serviceAccountKey: "{}"
   region: "eu01-1"
-  userData: |
-    #cloud-config
-    runcmd:
-      - echo "Machine bootstrapped"
 `, secretName, testNamespace)
 			createAndTrackResource(ctx, "secret", secretName, testNamespace, secretYAML)
