+++ test/e2e/e2e_userdata_test.go
@@ -34,11 +34,6 @@ stringData:
   projectId: "12345678-1234-1234-1234-123456789012"
   serviceAccountKey: "{}"
   region: "eu01-1"
-  networkId: "770e8400-e29b-41d4-a716-446655440000"
-  userData: |
-    #cloud-config
-    runcmd:
-      - echo "Base bootstrap from Secret"
 `, secretName, testNamespace)
 			createAndTrackResource(ctx, "secret", secretName, testNamespace, secretYAML)
